+++ src/Provider/ListenerCollection.php
@@ -8,6 +8,7 @@ use Closure;
 use InvalidArgumentException;
 use ReflectionFunction;
 use ReflectionNamedType;
+use ReflectionUnionType;
 
 /**
  * Listener collection stores listeners and is used to configure provider.
@@ -49,19 +50,21 @@ final class ListenerCollection
      * Any callable could be used be it a closure, invokable object or array referencing a class or object.
      *
      * @param callable $listener
-     * @param string $eventClassName
+     * @param string ...$eventClassName
      *
      * @return self
      */
-    public function add(callable $listener, string $eventClassName = ''): self
+    public function add(callable $listener, string ...$eventClassNames): self
     {
         $new = clone $this;
 
-        if ($eventClassName === '') {
-            $eventClassName = $this->getParameterType($listener);
+        if ($eventClassNames === []) {
+            $eventClassNames = $this->getParameterType($listener);
         }
 
-        $new->listeners[$eventClassName][] = $listener;
+        foreach ($eventClassNames as $eventClassName) {
+            $new->listeners[$eventClassName][] = $listener;
+        }
         return $new;
     }
 
@@ -70,18 +73,29 @@ final class ListenerCollection
      *
      * @param callable $callable The callable for which we want the parameter type.
      *
-     * @return string The interface the parameter is type hinted on.
+     * @return string[] Interfaces the parameter is type hinted on.
      */
-    private function getParameterType(callable $callable): string
+    private function getParameterType(callable $callable): array
     {
         $closure = new ReflectionFunction(Closure::fromCallable($callable));
         $params = $closure->getParameters();
 
         $reflectedType = isset($params[0]) ? $params[0]->getType() : null;
-        if ($reflectedType === null) {
-            throw new InvalidArgumentException('Listeners must declare an object type they can accept.');
+
+        if ($reflectedType instanceof ReflectionNamedType) {
+            return [$reflectedType->getName()];
         }
-        /** @var ReflectionNamedType $reflectedType */
-        return $reflectedType->getName();
+
+        /** @psalm-suppress UndefinedClass,TypeDoesNotContainType */
+        if ($reflectedType instanceof ReflectionUnionType) {
+            /** @var ReflectionNamedType[] */
+            $types = $reflectedType->getTypes();
+            return array_map(
+                static fn (ReflectionNamedType $type) => $type->getName(),
+                $types
+            );
+        }
+
+        throw new InvalidArgumentException('Listeners must declare an object type they can accept.');
     }
 }