+++ src/Provider/Provider.php
@@ -5,6 +5,7 @@ declare(strict_types=1);
 namespace Yiisoft\EventDispatcher\Provider;
 
 use Psr\EventDispatcher\ListenerProviderInterface;
+
 use function get_class;
 
 /**
@@ -32,7 +33,9 @@ final class Provider implements ListenerProviderInterface
     public function getListenersForEvent(object $event): iterable
     {
         yield from $this->listeners->getForEvents(get_class($event));
+        /** @psalm-suppress MixedArgument */
         yield from $this->listeners->getForEvents(...array_values(class_parents($event)));
+        /** @psalm-suppress MixedArgument */
         yield from $this->listeners->getForEvents(...array_values(class_implements($event)));
     }
 }