+++ CHANGELOG.md
@@ -1,9 +1,9 @@
-# Yii Dependency Injection Change Log
-
+# Yii Event Dispatcher Change Log
 
 ## 1.0.1 under development
 
-- no changes in this release.
+- New #63: Allow adding a single listener to multiple event class names at once in `ListenerCollection::add()` (vjik)
+- Bug #62: Support listeners for events defined with union types (vjik)
 
 ## 1.0.0 February 16, 2021